scrapers/requirements.txt
# Oasara Data Liberation Scrapers
# Phase 3: Pre-populate stories with real healthcare horror stories

# Core
python-dotenv==1.0.0
requests==2.31.0
httpx==0.27.0
aiohttp==3.9.1

# Reddit
praw==7.7.1
asyncpraw==7.7.1

# Twitter/X
tweepy==4.14.0
snscrape==0.7.0.20230622

# YouTube
yt-dlp==2024.1.0

# Web scraping
beautifulsoup4==4.12.2
lxml==5.1.0
playwright==1.41.0
selectolax==0.3.17

# Image processing
Pillow==10.2.0
pytesseract==0.3.10
opencv-python-headless==4.9.0.80

# Video/Audio
openai-whisper==20231117
ffmpeg-python==0.2.0

# AI extraction
anthropic==0.18.0
openai==1.12.0

# Storage
boto3==1.34.0  # S3-compatible (R2)
supabase==2.3.0

# NAS sync (system tool, install separately)
# rclone - https://rclone.org/install/

# Utilities
tqdm==4.66.1
tenacity==8.2.3
python-dateutil==2.8.2
uuid==1.30

